Created test_tags.py, added more tests, checked now defaults to None in filter_tasks_by_tags
